Move specification code to the Specification sub-module
You might think I would have done some sort of analyis of which
symbols to pull out into what modules, but you might be wrong.
Let's see how moving out the "specification" stuff works out!
