+++ hvega/hvega.cabal
@@ -123,6 +123,7 @@ library
   exposed-modules:     Graphics.Vega.VegaLite
                        Graphics.Vega.Tutorials.VegaLite
   other-modules:       Graphics.Vega.VegaLite.Core
+                       Graphics.Vega.VegaLite.Specification
   build-depends:       base >= 4.9 && < 5
                      , aeson >= 0.11 && < 1.5
                      , text == 1.2.*
